+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Added global styles for ACF admin layout.
+- Added few css classes for ACF -fields.
+
 ## [0.5.0] - 2017-09-08
 
 ### Changed