+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Moved Components::hook function to load on init instead of acf/init so we can use data that is set after init in fields
+
 ## [0.5.0] - 2017-09-08
 
 ### Changed