+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Moved Components::hook function to load on init instead of acf/init so we can use data that is set after init in fields
 
+## [0.5.1] - 2017-09-22
+
+### Added
+- Static components now have acf_fc_layout field as well
+
+### Changed
+- Static components don't have a prefix anymore as this would break component data functions
+- Changed clonable static component name from c to component name
+
+
 ## [0.5.0] - 2017-09-08
 
 ### Changed